DEVO-42 File restructuring
- Adds tasks directory and move task files
- Removes sc namespace from files and dags
- Removes old files that are no longer used
