+++ cob_datapipeline/catalog_move_alma_sftp_to_s3_dag.py
@@ -8,7 +8,7 @@ from airflow.contrib.hooks.sftp_hook import SFTPHook
 from airflow.operators.python_operator import PythonOperator
 from cob_datapipeline.operators.batch_sftp_to_s3_operator import BatchSFTPToS3Operator
 from cob_datapipeline.helpers import determine_most_recent_date
-from cob_datapipeline.sc_xml_parse import update_variables
+from cob_datapipeline.tasks.xml_parse import update_variables
 
 ALMA_SFTP_CONNECTION_ID = "ALMASFTP"
 S3_CONN_ID = "AIRFLOW_S3"
